Email Verification and Login changes
Login - return status 200 for errors so the front-end can provide the User with better error messages
Register - Add TokenCreationDateTime to the insert query.
VerifyEmail - Amend VerifyEmail Route to handle expiring tokens and send a new token if a user is attempting to verify with an expired token.
DatabaseSetupSQL - Added TokenCreationDateTime column to Account table